requirements.txt
pillow~=11.0.0
dearpygui~=1.10.0
numpy~=2.2.3